! Deleted ExamplePlayer and move all code to ExampleInfoDisplay and ExampleFireDebuff (for the sake of ExampleMods 'convention') * Added useful comments to ExampleInfoDisplay * Added some comments to ExampleFireDebuff
