+++ ExampleMod/Content/Buffs/ExampleFireDebuff.cs
@@ -1,12 +1,10 @@
-using ExampleMod.Common.Players;
-using ExampleMod.NPCs;
 using Terraria;
 using Terraria.ModLoader;
 
 namespace ExampleMod.Content.Buffs
 {
-	// Fire debuff is an example of a buff that causes constant loss of life.
-	// See ExamplePlayer.UpdateBadLifeRegen for more information.
+	// Fire debuff is an example of a buff that causes constant loss of life
+	// See ExampleFireDebuffPlayer.UpdateBadLifeRegen at the end of the file for more information
 	public class ExampleFireDebuff : ModBuff
 	{
 		public override void SetDefaults() {
@@ -20,7 +18,33 @@ namespace ExampleMod.Content.Buffs
 
 		// Allows you to make this buff give certain effects to the given player
 		public override void Update(Player player, ref int buffIndex) {
-			player.GetModPlayer<ExamplePlayer>().FireDebuff = true;
+			player.GetModPlayer<ExampleFireDebuffPlayer>().FireDebuff = true;
+		}
+	}
+
+	class ExampleFireDebuffPlayer: ModPlayer
+	{
+		// Flag checking when Fire Debuff should be activated
+		public bool FireDebuff;
+
+		public override void ResetEffects() {
+			FireDebuff = false;
+		}
+
+		// Allows you to give the player a negative life regeneration based on its state(for example, the "On Fire!" debuff makes the player take damage-over-time)
+		// This is typically done by setting player.lifeRegen to 0 if it is positive, setting player.lifeRegenTime to 0, and subtracting a number from player.lifeRegen
+		// The player will take damage at a rate of half the number you subtract per second
+		public override void UpdateBadLifeRegen() {
+			if (FireDebuff) {
+				// These lines zero out any positive lifeRegen. This is expected for all bad life regeneration effects
+				if (Player.lifeRegen > 0)
+					Player.lifeRegen = 0;
+				// Player.lifeRegenTime uses to increase the speed at which the player reaches its maximum natural life regeneration
+				// So we set it to 0, and while this debuff is active, it never reaches it
+				Player.lifeRegenTime = 0;
+				// lifeRegen is measured in 1/2 life per second. Therefore, this effect causes 8 life lost per second
+				Player.lifeRegen -= 16;
+			}
 		}
 	}
 }
